requirements.txt
aniso8601==9.0.1
apt-clone==0.2.1
apturl==0.5.2
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
ccsm==0.9.14.1
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
eyeD3==0.8.10
filelock==3.6.0
Flask==2.2.2
Flask-RESTful==0.3.9
grpcio==1.30.2
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
Mako==1.1.3
MarkupSafe==2.1.1
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
onboard==1.4.1
packaging==21.3
PAM==0.4.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.12.4
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
PyGObject==3.42.0
PyICU==2.8.1
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
python-apt==2.3.0+ubuntu2
python-debian===0.1.43ubuntu1
python-magic==0.4.24
python-xapp==2.2.2
python-xlib==0.29
pytz==2022.2.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
systemd-python==234
tldextract==3.1.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
Werkzeug==2.2.2
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
